+++ spring-server/src/main/java/ch/heigvd/gamification/database/dao/BadgeRuleRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 /**
  * Created by lux on 10.12.16.
  */
-public interface BadgeRuleRepository extends CrudRepository<BadgeRule, Integer> {
+public interface BadgeRuleRepository extends CrudRepository<BadgeRule, Long> {
     List<BadgeRule> findAll();
     List<BadgeRule> findByApplication(Application application);
 }
